requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.13.3
blis==1.2.0
bs4==0.0.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
jericho @ file:///home/abd/Developer/Python/ai-agent-SOAR/jericho
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langcodes==3.5.0
langgraph==0.2.59
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.46
langsmith==0.2.3
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nautilus-open-any-terminal==0.6.0
numpy==1.26.4
openai==1.57.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.8.3
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.1
starlette==0.41.3
tenacity==9.0.0
termcolor==2.5.0
thinc==8.3.4
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
websockets==14.1
wrapt==1.17.2
yarl==1.18.3
